Add files via upload
An application has been developed where users can take notes. The application will display notes with different appearance on the screen according to the user's preferences.

The application supports 3 different types: list notes, tabular notes and regular plain text notes. If they want to keep his notes in a list, they can choose to be ordered or unordered.

When viewing notes when sequential is selected 1. 2. 3. or a. b. c. It should specify which of the formats it will be in. 1. 2. 3. For 1, a. b. c. You can specify it as a.

When it is selected unordered, it should specify which character it will be listed with. Valid characters are *, -, . is in the form. Then, they can create and view the list note by entering the information they want.

If they want to keep his notes in tabular form, they must first enter the header information of the columns. Then, line by line, which data belongs to which column, it will enter its data accordingly. Can create and view the table

If they want to keep his notes in plain text, they can enter the title first. Then they can view his paragraphs after creating the note.

The users can select the desired text color for all note types.